content-security-policy/to-port-from-blink/1.1/scriptnonce-basic-blocked-expected.txt

ALERT: PASS (1/2)
ALERT: PASS (2/2)
CONSOLE ERROR: line 13: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'nonce-noncynonce'". Either the 'unsafe-inline' keyword, a hash ('sha256-1RpMm8CnZ3w3FNN-LbHHalT1URf5MlDZCUee2MAF0_U='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 16: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'nonce-noncynonce'". Either the 'unsafe-inline' keyword, a hash ('sha256-aA5X7j5nr6U308mrny9CeIP0WyoPjaNn_0piYZqVL7I='), or a nonce ('nonce-...') is required to enable inline execution.

CONSOLE ERROR: line 19: Refused to execute inline script because it violates the following Content Security Policy directive: "script-src 'nonce-noncynonce'". Either the 'unsafe-inline' keyword, a hash ('sha256-DnnF--Kbf_s2EyTBk12Rw2Ouq__L7GY5pcHSc6IZe6U='), or a nonce ('nonce-...') is required to enable inline execution.

This tests the effect of a valid script-nonce value. It passes if three console warnings are visible, and the two PASS alerts are executed.
